# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Amrutam Telemedicine API",
    "description": "Production-grade telemedicine backend: auth/MFA, doctor availability & booking (idempotent, saga-driven), consultations, prescriptions, payments, admin analytics.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      }
    }
  },
  "paths": {
...
    "/prescriptions": {
      "post": {
        "summary": "Doctor issues a prescription for a consultation (encrypted at rest)",
        "tags": [
          "prescriptions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "consultationId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "content": {
                    "type": "object",
                    "properties": {
                      "medications": {
                        "minItems": 1,
                        "maxItems": 50,
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "minLength": 1,
                              "maxLength": 200,
                              "type": "string"
                            },
                            "dosage": {
                              "minLength": 1,
                              "maxLength": 100,
                              "type": "string"
                            },
                            "frequency": {
                              "minLength": 1,
                              "maxLength": 100,
                              "type": "string"
                            },
                            "durationDays": {
                              "minimum": 1,
                              "maximum": 365,
                              "type": "integer"
                            },
                            "notes": {
                              "maxLength": 500,
                              "type": "string"
                            }
                          },
                          "required": [
                            "name",
                            "dosage",
                            "frequency",
                            "durationDays"
                          ]
                        }
                      },
                      "advice": {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      "followUpInDays": {
                        "minimum": 1,
                        "maximum": 365,
                        "type": "integer"
                      }
                    },
                    "required": [
                      "medications"
                    ]
                  },
                  "supersedesId": {
                    "format": "uuid",
                    "type": "string"
                  }
                },
                "required": [
                  "consultationId",
                  "content"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "minLength": 8,
              "maxLength": 200,
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true,
            "description": "Client-generated unique key that makes this write idempotent."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "consultationId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "doctorId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "patientId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "content": {
                      "type": "object",
                      "properties": {
                        "medications": {
                          "minItems": 1,
                          "maxItems": 50,
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "minLength": 1,
                                "maxLength": 200,
                                "type": "string"
                              },
                              "dosage": {
                                "minLength": 1,
                                "maxLength": 100,
                                "type": "string"
                              },
                              "frequency": {
                                "minLength": 1,
                                "maxLength": 100,
                                "type": "string"
                              },
                              "durationDays": {
                                "minimum": 1,
                                "maximum": 365,
                                "type": "integer"
                              },
                              "notes": {
                                "maxLength": 500,
                                "type": "string"
                              }
                            },
                            "required": [
                              "name",
                              "dosage",
                              "frequency",
                              "durationDays"
                            ]
                          }
                        },
                        "advice": {
                          "maxLength": 2000,
                          "type": "string"
                        },
                        "followUpInDays": {
                          "minimum": 1,
                          "maximum": 365,
                          "type": "integer"
                        }
                      },
                      "required": [
                        "medications"
                      ]
                    },
                    "issuedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "supersedesId": {
                      "anyOf": [
                        {
                          "format": "uuid",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "consultationId",
                    "doctorId",
                    "patientId",
                    "content",
                    "issuedAt",
                    "supersedesId"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List my prescriptions (patient)",
        "tags": [
          "prescriptions"
        ],
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "consultationId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "doctorId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "patientId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "content": {
                            "type": "object",
                            "properties": {
                              "medications": {
                                "minItems": 1,
                                "maxItems": 50,
                                "type": "array",
                                "items": {
                                  "type": "object",
                                  "properties": {
                                    "name": {
                                      "minLength": 1,
                                      "maxLength": 200,
                                      "type": "string"
                                    },
                                    "dosage": {
                                      "minLength": 1,
                                      "maxLength": 100,
                                      "type": "string"
                                    },
                                    "frequency": {
                                      "minLength": 1,
                                      "maxLength": 100,
                                      "type": "string"
                                    },
                                    "durationDays": {
                                      "minimum": 1,
                                      "maximum": 365,
                                      "type": "integer"
                                    },
                                    "notes": {
                                      "maxLength": 500,
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "name",
                                    "dosage",
                                    "frequency",
                                    "durationDays"
                                  ]
                                }
                              },
                              "advice": {
                                "maxLength": 2000,
                                "type": "string"
                              },
                              "followUpInDays": {
                                "minimum": 1,
                                "maximum": 365,
                                "type": "integer"
                              }
                            },
                            "required": [
                              "medications"
                            ]
                          },
                          "issuedAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "supersedesId": {
                            "anyOf": [
                              {
                                "format": "uuid",
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "consultationId",
                          "doctorId",
                          "patientId",
                          "content",
                          "issuedAt",
                          "supersedesId"
                        ]
                      }
                    },
                    "nextCursor": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/prescriptions/{id}": {
      "get": {
        "summary": "Get a prescription (patient owner or issuing doctor)",
        "tags": [
          "prescriptions"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "consultationId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "doctorId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "patientId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "content": {
                      "type": "object",
                      "properties": {
                        "medications": {
                          "minItems": 1,
                          "maxItems": 50,
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "minLength": 1,
                                "maxLength": 200,
                                "type": "string"
                              },
                              "dosage": {
                                "minLength": 1,
                                "maxLength": 100,
                                "type": "string"
                              },
                              "frequency": {
                                "minLength": 1,
                                "maxLength": 100,
                                "type": "string"
                              },
                              "durationDays": {
                                "minimum": 1,
                                "maximum": 365,
                                "type": "integer"
                              },
                              "notes": {
                                "maxLength": 500,
                                "type": "string"
                              }
                            },
                            "required": [
                              "name",
                              "dosage",
                              "frequency",
                              "durationDays"
                            ]
                          }
                        },
                        "advice": {
                          "maxLength": 2000,
                          "type": "string"
                        },
                        "followUpInDays": {
                          "minimum": 1,
                          "maximum": 365,
                          "type": "integer"
                        }
                      },
                      "required": [
                        "medications"
                      ]
                    },
                    "issuedAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "supersedesId": {
                      "anyOf": [
                        {
                          "format": "uuid",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "consultationId",
                    "doctorId",
                    "patientId",
                    "content",
                    "issuedAt",
                    "supersedesId"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}